+++ src/core/internalTypes.ts
@@ -148,7 +148,7 @@ export interface IParseCoreOptions {
     isWithTiming: boolean // (Requires isIncludeMeta) Include timing data of the different phases in meta data, when isIncludeMeta.
     isKeepUndefinedInMeta: boolean // (Requires isIncludeMeta) If true, keeps properties with undefined values in the returned meta data, when isIncludeMeta.
     // isAvoidWarningsInConsole: boolean // (!) Dup of quiet! - Suppress warnings (make quiet) in console (does not affect warnings in meta data).
-    isQuiet: boolean // Dup of suppressWarnings! Reduce output (show only errors, does not affect warnings in meta data).
+    isQuiet: boolean // Suppress-warnings, reduce output (show only errors, does not affect warnings in meta data).
     isSilent: boolean // Suppress all output (even errors, exit code only).
     isThrowOnError: boolean // Will throw on first parse error encountered.
 }