+++ src/core/options/defaultParserOptions.ts
@@ -1,15 +1,10 @@
 import { ParseOptions } from '../../types'
 import { TParserMode } from '../internalTypes'
 
-// export const getDefaultUserOptions = (mode: TParserMode): ParseOptions =>
-// export const getDefaultUserOptions = (mode: TParserMode) =>
-//     mode === 'strict' ? DEFAULT_STRICT_OPTS : DEFAULT_LENIENT_OPTS
-
 export const getDefaultUserOptions = (mode: TParserMode) =>
     mode === 'strict' ? { ...DEFAULT_STRICT_OPTS } : { ...DEFAULT_LENIENT_OPTS }
 
 export type TNormalizedUserOptions = Required<
-    // type TNormalizedUserOptions = Required<
     Pick<
         ParseOptions,
         | 'strictMode'
@@ -18,7 +13,6 @@ export type TNormalizedUserOptions = Required<
         | 'includeDiagnostics'
         | 'includeTiming'
         | 'preserveUndefinedInMeta'
-        // | 'suppressWarnings' // Delete, dup of quiet!
         | 'onDuplicateKey'
         | 'requireDocTerminator'
         | 'treatEmptyValueAsNull'
@@ -28,7 +22,7 @@ export type TNormalizedUserOptions = Required<
     >
 >
 
-// Base (mode-agnostic) defaults
+// Base (mode-agnostic) defaults.
 const BASE_DEFAULTS: TNormalizedUserOptions = {
     strictMode: false,
     failLevel: 'auto',
@@ -36,7 +30,6 @@ const BASE_DEFAULTS: TNormalizedUserOptions = {
     includeDiagnostics: false,
     includeTiming: false,
     preserveUndefinedInMeta: false,
-    // suppressWarnings: false, // Suppress warnings in console (does not affect warnings in meta data).
     onDuplicateKey: 'error',
     requireDocTerminator: 'optional',
     treatEmptyValueAsNull: 'allow-with-warning',
@@ -50,7 +43,6 @@ const DEFAULT_LENIENT_OPTS: TNormalizedUserOptions = {
     ...BASE_DEFAULTS,
     strictMode: false,
     failLevel: 'ignore-errors',
-    // suppressWarnings: false, // Suppress warnings in console (does not affect warnings in meta data).
     // Below are options for pure rules:
     onDuplicateKey: 'warn-and-keep-first',
     requireDocTerminator: 'optional',
@@ -61,7 +53,6 @@ const DEFAULT_STRICT_OPTS: TNormalizedUserOptions = {
     ...BASE_DEFAULTS,
     strictMode: true,
     failLevel: 'errors',
-    // suppressWarnings: false, // Suppress warnings in console (does not affect warnings in meta data).
     // Below are options for pure rules:
     onDuplicateKey: 'error',
     requireDocTerminator: 'optional',